Update font-face declarations to use WOFF2 font formats
This commit updates the `@font-face` declarations within the project's CSS to reference the WOFF2 versions of the 'Snell Roundhand' font family. Previously, the CSS was pointing to TTF font files for normal, bold, and black weights. This change is made to leverage the benefits of the WOFF2 format, which includes improved compression and faster load times, enhancing the website's performance and user experience.

Changes include:
- Removing references to the TTF font files.
- Adding `src` attributes to exclusively use `.woff2` font files for the 'Snell Roundhand' font family across normal, bold, and black weights.

This update is part of ongoing efforts to optimize the website's assets for speed and efficiency, ensuring a smoother and more responsive experience for all users.