Update favicon path in admin section
This commit updates the path to the favicon for the admin section of the website. The previous relative path was not correctly pointing to the favicon file, resulting in the icon not displaying properly in browser tabs for admin pages.

- Changed the favicon link from `<link rel="icon" type="image/png" href="img/favicon-64.png">`
- To `<link rel="icon" type="image/png" href="../img/favicon-64.png">`

This adjustment corrects the path relative to the admin directory's structure, ensuring the favicon is consistently displayed across the site.